libs/timer: Move into internal
