+++ abci/client/socket_client.go
@@ -14,7 +14,7 @@ import (
 	"github.com/cometbft/cometbft/abci/types"
 	cmtnet "github.com/cometbft/cometbft/internal/net"
 	"github.com/cometbft/cometbft/internal/service"
-	"github.com/cometbft/cometbft/libs/timer"
+	"github.com/cometbft/cometbft/internal/timer"
 )
 
 const (
